📦 chore(package.json): add @nestjs/schedule and node-binance-api dependencies
The @nestjs/schedule package is added to enable scheduling tasks within the NestJS application. This will be used to schedule periodic tasks in the server.

The node-binance-api package is added to provide an API client for interacting with the Binance cryptocurrency exchange. This will be used to fetch data from Binance and perform trading operations.